GVT-1848: Move files around
