+++ ui/src/map/layers/switch/switch-layer.ts
@@ -7,10 +7,10 @@ import { MapTile, OptionalShownItems } from 'map/map-model';
 import { Selection } from 'selection/selection-model';
 import { LayoutSwitch, LayoutSwitchId } from 'track-layout/track-layout-model';
 import { getSwitchesByTile } from 'track-layout/layout-switch-api';
-import { getMatchingSwitches } from 'map/layers/layer-utils';
-import { MapLayer, SearchItemsOptions } from 'map/layers/layer-model';
-import * as Limits from 'map/layers/layer-visibility-limits';
-import { createFeatures } from 'map/layers/switch-layer-utils';
+import { getMatchingSwitches } from 'map/layers/utils/layer-utils';
+import { MapLayer, SearchItemsOptions } from 'map/layers/utils/layer-model';
+import * as Limits from 'map/layers/utils/layer-visibility-limits';
+import { createFeatures } from 'map/layers/switch/switch-layer-utils';
 import { PublishType, TimeStamp } from 'common/common-model';
 import { getSwitchStructures } from 'common/common-api';
 import { ChangeTimes } from 'common/common-slice';
