+++ ui/src/map/layers/switch/switch-manual-linking-layer.ts
@@ -3,11 +3,11 @@ import { Point as OlPoint, Polygon } from 'ol/geom';
 import { Vector as VectorLayer } from 'ol/layer';
 import { Vector as VectorSource } from 'ol/source';
 import { MapTile } from 'map/map-model';
-import { LayerItemSearchResult, MapLayer, SearchItemsOptions } from 'map/layers/layer-model';
+import { LayerItemSearchResult, MapLayer, SearchItemsOptions } from 'map/layers/utils/layer-model';
 import { LocationTrackEndpoint } from 'linking/linking-model';
-import { MANUAL_SWITCH_LINKING_ENDPOINT_SELECTION_RESOLUTION } from 'map/layers/layer-visibility-limits';
+import { MANUAL_SWITCH_LINKING_ENDPOINT_SELECTION_RESOLUTION } from 'map/layers/utils/layer-visibility-limits';
 import { flatten } from 'utils/array-utils';
-import { getMatchingEntities, MatchOptions, pointToCoords } from 'map/layers/layer-utils';
+import { getMatchingEntities, MatchOptions, pointToCoords } from 'map/layers/utils/layer-utils';
 import { endPointStyle } from 'map/layers/linking-layer';
 import { PublishType } from 'common/common-model';
 import { getLocationTrackEndpointsByTile } from 'track-layout/layout-location-track-api';
