+++ ui/src/map/layers/utils/layer-model.ts
