+++ ui/src/map/layers/utils/layer-visibility-limits.ts
@@ -1,4 +1,4 @@
-import { MapLayerName } from "map/map-model";
+import { MapLayerName } from 'map/map-model';
 // Alignments
 export const ALL_ALIGNMENTS = 10.0;
 export const LINKING_DOTS = 0.19;
