+++ ui/src/map/layers/utils/rendering.ts
