+++ ui/src/map/settings-menu/map-layer-menu.scss
