+++ ui/src/map/tools/select-tool.ts
@@ -1,7 +1,7 @@
 import OlMap from 'ol/Map';
 import { MapTool, MapToolActivateOptions } from './tool-model';
 import { getDefaultHitArea, searchItemsFromLayers } from 'map/tools/tool-utils';
-import { MapLayer } from 'map/layers/layer-model';
+import { MapLayer } from 'map/layers/utils/layer-model';
 
 export const selectTool: MapTool = {
     activate: (map: OlMap, layers: MapLayer[], options: MapToolActivateOptions) => {
